move equal drift results into common object to support osp
